rng_operator/training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.14.0
onnxruntime>=1.15.0
tqdm
pillow
numpy
opencv-python-headless
onnxscript>=0.1.0
